Add files via upload
With the rapid evolution of technology, the threat landscape for cybersecurity has become increasingly sophisticated, with malware being a pervasive and persistent menace. This study presents a comprehensive comparative analysis of various machine learning (ML) techniques for malware detection and mitigation. The primary objective is to evaluate the efficacy of different ML algorithms in identifying and combating malware threats.